feat(wintertc): migrate console from boa_runtime
This Pull Request is part of the TC55 migration tracked in #4988, and
follows the move + re-export pattern established by the base64 migration
(#5418).

**It changes the following:**

- **Moves the `console` module** (`mod.rs`, `table.rs`, and `tests.rs`)
into `boa_wintertc`. The implementation is **unchanged**: `table.rs` is
byte for byte identical, `tests.rs` differs only in one import path
(`Logger`/`NullLogger` now come from the console module rather than the
`boa_runtime` crate root), and `mod.rs` only gains a short TC55 status
note and a thin `register` wrapper (see below).
- **Deletes `core/runtime/src/console/`** and **re-exports the module
from `boa_runtime`**: `boa_runtime::console` is now `#[doc(inline)] pub
use boa_wintertc::console;`. The crate-root re-export `pub use
console::{Console, ConsoleState, DefaultLogger, Logger, NullLogger}`
keeps resolving through it, so the public API is unchanged.
- **Delegates the extension**: `ConsoleExtension` now calls
`boa_wintertc::console::Console::register_with_logger`. It stays in the
default extension set, so there is no behaviour change for existing
users.
- **Adds a `register` entry point** to `boa_wintertc::console` so it
plugs into `boa_wintertc::register` like the other modules. It installs
`console` with the `DefaultLogger`; `register_with_logger` remains the
way to supply a custom logger.
- **Moves the `comfy-table` dependency** from `boa_runtime` to
`boa_wintertc`, as the moved module was its only user, and adds `boa_gc`
and `rustc-hash` to `boa_wintertc` (both already used by the console
implementation).
- **Allows `needless_raw_string_hashes` in test builds** of
`boa_wintertc`, matching `boa_runtime`; the console tests use raw JS
strings kept copy-pastable.

**Two adjustments in `boa_runtime` where console's tests were shared:**

- `microtask/tests.rs` used `console::tests::RecordingLogger`, a
`#[cfg(test)]` helper that has now left the crate. It defines its own
small `RecordingLogger` locally instead. This is temporary: the helper
goes away when `microtask` migrates.
- `run_test_actions` (the test harness variant that registers a console)
was previously kept alive by the always-compiled console tests. Its only
remaining callers are the feature-gated `fetch`, `url`, and `abort`
tests, so it is now `#[allow(unused)]` to stay warning-free under
`--no-default-features`.

**Line counts (before and after, in both locations):**

| File | `boa_runtime` (before) | `boa_wintertc` before | `boa_wintertc` after |
| --- | --- | --- | --- |
| `console/mod.rs` | 1008 (deleted) | 23 (stub) | 1027 |
| `console/table.rs` | 214 (deleted) | 0 (did not exist) | 214 |
| `console/tests.rs` | 853 (deleted) | 0 (did not exist) | 852 |

`mod.rs` gains 19 lines: the TC55 status doc note and the `register`
wrapper. `table.rs` is identical; `tests.rs` is one line shorter after
merging two imports into one.

**Testing:**

- `cargo test -p boa_wintertc`: all console unit tests pass (46 total
including base64 and the doctest).
- `cargo test -p boa_runtime` and `--all-features`: pass, with console
now exercised through the re-export.
- `cargo clippy -p boa_wintertc` and `-p boa_runtime`, each with
`--all-features --all-targets` and `--no-default-features`: clean.
- `cargo fmt --all -- --check` and `cargo doc --no-deps`: clean.
- Verified `console.log`/`info`/`warn`/`error`, `group` indentation,
`console.table`, `count`, and `dir` end to end through `boa_cli`.

**Out of scope:**

- **WPT** is unchanged: the runner targets `boa_runtime`, which reaches
this implementation through the re-export.
- **Test262** does not apply, as `console` is a Web API rather than
ECMAScript.
