+++ core/runtime/Cargo.toml
@@ -21,7 +21,6 @@ futures = "0.3.32"
 futures-lite.workspace = true
 http = { workspace = true, optional = true }
 reqwest = { workspace = true, optional = true }
-comfy-table.workspace = true
 rustc-hash = { workspace = true, features = ["std"] }
 serde_json = { workspace = true, optional = true }
 url = { workspace = true, optional = true }
