+++ core/wintertc/Cargo.toml
@@ -12,7 +12,9 @@ rust-version.workspace = true
 
 [dependencies]
 boa_engine.workspace = true
+boa_gc.workspace = true
 base64.workspace = true
+comfy-table.workspace = true
 rustc-hash = { workspace = true, features = ["std"] }
 
 [dev-dependencies]
