+++ core/wintertc/src/console/table.rs
